Extracted SignUpForm into component
