+++ app/signup/signup-form.tsx
@@ -1,0 +1,85 @@
+"use client"; 
+import { useState } from "react";
+
+
+export default function SignUpForm() {
+  const [formData, setFormData] = useState({
+    name: "",
+    email: "",
+    password: "",
+  });
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setFormData({ ...formData, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    console.log("Form Submitted:", formData);
+  };
+
+  return (
+    <div className="bg-white shadow-md rounded-lg p-8 w-full max-w-md">
+      <h2 className="text-2xl font-bold text-center text-gray-800 mb-6">
+        Sign Up
+      </h2>
+      <form onSubmit={handleSubmit} className="space-y-4">
+        {/* Name Field */}
+        <div>
+          <label className="block text-gray-700 text-sm font-medium mb-1">
+            Name
+          </label>
+          <input
+            type="text"
+            name="name"
+            value={formData.name}
+            onChange={handleChange}
+            required
+            className="w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
+            placeholder="Enter your name"
+          />
+        </div>
+
+        {/* Email Field */}
+        <div>
+          <label className="block text-gray-700 text-sm font-medium mb-1">
+            Email
+          </label>
+          <input
+            type="email"
+            name="email"
+            value={formData.email}
+            onChange={handleChange}
+            required
+            className="w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
+            placeholder="Enter your email"
+          />
+        </div>
+
+        {/* Password Field */}
+        <div>
+          <label className="block text-gray-700 text-sm font-medium mb-1">
+            Password
+          </label>
+          <input
+            type="password"
+            name="password"
+            value={formData.password}
+            onChange={handleChange}
+            required
+            className="w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
+            placeholder="Enter your password"
+          />
+        </div>
+
+        {/* Submit Button */}
+        <button
+          type="submit"
+          className="w-full bg-blue-500 text-white py-2 rounded-lg hover:bg-blue-600 transition duration-300"
+        >
+          Sign Up
+        </button>
+      </form>
+    </div>
+  );
+}
